working of Ai tourism chatbot
The above observed results indicate the initial step of building a Rule-based chatbot based on Lstm model. The Rule-Based Chatbot code is observed to be running Successfully based and get Responses based of  Data set based on user can be inculed in the Data set. 
Partically tested  a Data Set of Cinema Dialogues getting Successfully of Accuracy 89%  by Model Rms Prop and  Adam Algorithm is Getting 92%.
But we created new Data set called Tourism Guide to get response But Iam Failed of getting 100% Accuracy so technically Data Set is over fitting. But it is giving responses based on user FAQ’s 
By the above problem is to solving  using AI tools of managing Dataset and avoid using Dataset(csv) it will not use for Future scope to use a Database to handle user queries.
Note: Use AI tools to be supported by Tensorflow and keras Software by Adjusting the code.